# apps/control-plane/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenSesame Control Plane",
    "version": "0.1.0",
    "description": "Identity, claims, provisional principals, and agent registration APIs"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8788"
    }
  ],
  "paths": {
...
    "/v1/organizations": {
      "get": {
        "summary": "List organizations the caller belongs to, including their role",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "provisionalCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Organizations and caller roles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["organizations"],
                  "properties": {
                    "organizations": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Organization"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create an organization; the creator becomes owner",
        "description": "Cookie-authenticated browser mutations also require an allowed Origin header.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "provisionalCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganization"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          },
          "400": {
            "description": "Invalid organization body"
          },
          "401": {
            "description": "Authentication or cookie Origin required"
          },
          "403": {
            "description": "Verified identity required"
          },
          "409": {
            "description": "Slug already exists"
          }
        }
      }
    },
...
    "/v1/projects": {
      "get": {
        "summary": "List projects the caller can use, provisioning the personal default on first touch",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "provisionalCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Projects and caller roles, personal project first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["projects"],
                  "properties": {
                    "projects": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Project"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create a project; the creator becomes owner",
        "description": "Cookie-authenticated browser mutations also require an allowed Origin header.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "provisionalCookie": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProject"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project body"
          },
          "401": {
            "description": "Authentication or cookie Origin required"
          },
          "403": {
            "description": "Verified identity or quota required"
          },
          "404": {
            "description": "Organization not found"
          },
          "409": {
            "description": "Slug already exists"
          }
        }
      }
    },
...
    "/v1/claims": {
      "post": {
        "summary": "Create claim",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Missing principal bearer"
          },
          "403": {
            "description": "Quota exceeded"
          }
        }
      }
    },
...
    "/v1/agents": {
      "post": {
        "summary": "Register provisional agent",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Missing principal bearer"
          },
          "403": {
            "description": "Quota exceeded"
          }
        }
      }
    },
...
    "/v1/authorization-requests": {
      "get": {
        "summary": "The caller's own authorization-request inbox",
        "description": "Only ever the caller's own requests; there is no route that lists somebody else's.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "provisionalCookie": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AuthorizationRequestStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorization requests awaiting or settled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["requests"],
                  "properties": {
                    "requests": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AuthorizationRequest"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Ask a principal to authorize something",
        "description": "Addressed by the approver's inbox handle, never by a principal id: an unverifiable handle and a handle for a principal that must not be asked both answer 404, so nothing here confirms that an id exists. Honours Idempotency-Key.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "provisionalCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthorizationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Request created and waiting",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationRequest"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "not_found: the approver handle does not verify, or names a principal that must not be asked"
          }
        }
      }
    },
...
    "/v1/interactions": {
      "get": {
        "summary": "The approver's interaction inbox",
        "description": "A survey, not a presentation: listing does not move anything to awaiting_approval.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "provisionalCookie": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/InteractionStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Interactions awaiting this approver",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["interactions"],
                  "properties": {
                    "interactions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/InteractionDetail"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Raise a cross-device interaction over a ceremony",
        "description": "The server derives the binding message from the authorization details and computes the request digest over both; neither is taken from the body. Honours Idempotency-Key. Cookie-authenticated browser mutations also require an allowed Origin header.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "provisionalCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInteraction"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Interaction created; the reference is safe to print",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionCreated"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "interaction_not_found: the approver handle does not verify, or names a principal that must not be asked"
          },
          "409": {
            "description": "interaction_already_live: one live interaction per ceremony"
          },
          "422": {
            "description": "unsupported_kind, or an authorization detail that carries card data"
          }
        }
      }
    },
...
    "/v1/support": {
      "head": {
        "summary": "Check the authenticated same-origin support session",
        "responses": {
          "204": {
            "description": "Active session; X-OpenSesame-Support-Session: active"
          },
          "401": {
            "description": "Validated session required"
          },
          "404": {
            "description": "Remote support disabled"
          }
        }
      },
      "post": {
        "summary": "Send a preview-approved, redacted support request",
        "description": "Requires an active session and exact same-origin POST. Remote support is optional and disabled without server configuration.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "version",
                  "question",
                  "pageId",
                  "route",
                  "featureIds"
                ],
                "properties": {
                  "version": {
                    "const": 2
                  },
                  "question": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2000
                  },
                  "pageId": {
                    "type": "string",
                    "maxLength": 64
                  },
                  "route": {
                    "type": "string",
                    "maxLength": 65
                  },
                  "featureIds": {
                    "type": "array",
                    "maxItems": 32,
                    "items": {
                      "type": "string",
                      "maxLength": 64
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bounded AG-UI SSE response"
          },
          "400": {
            "description": "Invalid closed payload"
          },
          "401": {
            "description": "Validated session required"
          },
          "403": {
            "description": "Origin refused"
          },
          "413": {
            "description": "Payload exceeds 8KiB"
          },
          "502": {
            "description": "Upstream unavailable"
          }
        }
      }
    },
...
    "/v1/notification-channels": {
      "get": {
        "summary": "Channel capabilities, and whether an adapter is configured",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "provisionalCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Capability catalogue"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
...
    "/v1/notification-preferences": {
      "get": {
        "summary": "Where the caller prefers to be interrupted",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "provisionalCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Preferences by notification class"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      },
      "put": {
        "summary": "Record a preference",
        "description": "Stores the wish only. The intersection with policy, live bindings and configured adapters happens at routing time, so nothing here can widen what an operator allows.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "provisionalCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Stored"
          },
          "400": {
            "description": "Invalid preference"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateOrganization": {
        "type": "object",
        "additionalProperties": false,
        "required": ["slug", "displayName"],
        "properties": {
          "slug": {
            "type": "string",
            "minLength": 2,
            "maxLength": 64,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "ssoIssuer": {
            "type": "string",
            "format": "uri"
          },
          "samlIssuer": {
            "type": "string",
            "format": "uri"
          }
        }
      },
...
      "Organization": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "slug",
          "displayName",
          "state",
          "role",
          "createdBy",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/OrganizationState"
          },
          "role": {
            "$ref": "#/components/schemas/OrganizationRole"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "ssoIssuer": {
            "type": "string",
            "format": "uri"
          },
          "ssoClientId": {
            "type": "string"
          },
          "ssoClientSecretConfigured": {
            "type": "boolean"
          },
          "samlIssuer": {
            "type": "string",
            "format": "uri"
          }
        }
      },
...
      "CreateProject": {
        "type": "object",
        "additionalProperties": false,
        "required": ["displayName"],
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "slug": {
            "type": "string",
            "minLength": 2,
            "maxLength": 64,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "organizationId": {
            "type": "string"
          }
        }
      },
      "Project": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "kind",
          "slug",
          "displayName",
          "state",
          "role",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ProjectKind"
          },
          "slug": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ProjectState"
          },
          "role": {
            "$ref": "#/components/schemas/ProjectRole"
          },
          "ownerPrincipalId": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "AuthorizationRequestStatus": {
        "type": "string",
        "enum": ["pending", "approved", "denied", "expired", "cancelled"]
      },
...
      "CreateAuthorizationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["approverRef", "authorizationDetails", "bindingMessage"],
        "properties": {
          "approverRef": {
            "type": "string",
            "minLength": 8,
            "maxLength": 256
          },
          "authorizationDetails": {
            "type": "array",
            "minItems": 1,
            "maxItems": 32,
            "items": {
              "$ref": "#/components/schemas/AuthorizationDetail"
            }
          },
          "bindingMessage": {
            "type": "string",
            "minLength": 4,
            "maxLength": 120
          },
          "connectionId": {
            "type": "string"
          },
          "delegationId": {
            "type": "string"
          },
          "ttlSeconds": {
            "type": "integer",
            "minimum": 30,
            "maximum": 3600
          }
        }
      },
      "AuthorizationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "authReqId",
          "status",
          "bindingMessage",
          "requestDigest",
          "authorizationDetails",
          "expiresAt",
          "intervalSeconds"
        ],
        "properties": {
          "authReqId": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AuthorizationRequestStatus"
          },
          "bindingMessage": {
            "type": "string"
          },
          "requestDigest": {
            "type": "string",
            "description": "The canonical hash of exactly what is being consented to. An executor recomputes it before running anything."
          },
          "authorizationDetails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorizationDetail"
            }
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "intervalSeconds": {
            "type": "integer"
          },
          "connectionId": {
            "type": "string"
          },
          "delegationId": {
            "type": "string"
          },
          "decidedAt": {
            "type": "string",
            "format": "date-time"
          },
          "decidedByKind": {
            "$ref": "#/components/schemas/ApprovalDecidedByKind"
          }
        }
      },
...
      "InteractionStatus": {
        "type": "string",
        "enum": [
          "pending",
          "presented",
          "awaiting_approval",
          "approved",
          "denied",
          "consumed",
          "expired",
          "revoked"
        ]
      },
...
      "InteractionCreated": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "ref",
          "url",
          "requestDigest",
          "bindingMessage",
          "expiresAt",
          "status"
        ],
        "properties": {
          "ref": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "The canonical HTTPS short link, <publicUrl>/i/<ref>. Never a custom scheme."
          },
          "requestDigest": {
            "type": "string"
          },
          "bindingMessage": {
            "type": "string",
            "description": "Derived from the authorization details, never supplied by the requester."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/InteractionStatus"
          }
        }
      },
...
      "InteractionDetail": {
        "type": "object",
        "additionalProperties": false,
        "description": "The authenticated approver's view. The approval proof and the fronted ceremony's subjectId are absent by construction.",
        "required": [
          "kind",
          "status",
          "expiresAt",
          "requiresApprover",
          "id",
          "authorizationDetails",
          "createdAt"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/InteractionKind"
          },
          "status": {
            "$ref": "#/components/schemas/InteractionStatus"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "requiresApprover": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "requesterRef": {
            "type": "string"
          },
          "bindingMessage": {
            "type": "string"
          },
          "requestDigest": {
            "type": "string"
          },
          "authorizationDetails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorizationDetail"
            }
          },
          "resourceRef": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "decidedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "authenticationApiSecret": {
        "type": "http",
        "scheme": "bearer",
        "description": "Shown-once osa_ backend API secret"
      },
      "provisionalCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "os_provisional"
      }
    }
  }
}